% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.funLevels.R
\name{.check_args}
\alias{.check_args}
\title{Validate Arguments}
\usage{
.check_args(x, fun_message)
}
\arguments{
\item{x}{A character vector. Must not be \code{NULL}.}

\item{fun_message}{A character vector. Must not be \code{NULL}.}
}
\value{
This function does not return a value. It throws an error if any of the conditions are not met.
}
\description{
This function checks whether the provided arguments meet the expected conditions.
It ensures that \code{x} and \code{fun_message} are not \code{NULL} and that both are character vectors.
}
