% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.funLevels.R
\name{ds.funLevels}
\alias{ds.funLevels}
\title{Returns levels of a factor variable with fun message}
\usage{
ds.funLevels(x = NULL, fun_message = NULL, datasources = NULL)
}
\arguments{
\item{x}{a character string providing the name of the input data frame or matrix.}

\item{fun_message}{a character string with the motivational message}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.
If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
\code{ds.funLevels} returns the levels prefixed by a fun message
}
\description{
Retrieves the number of columns in a dataset and provides a motivational message
}
\examples{
\dontrun{
ds.funLevels(
  df.name = "iris$Species",
  fun_message = "This is a fun message ,",
  dataources = conns
)
}
}
\author{
Tim Cadman
}
